Fixed a memory leak regarding PrerequisiteLists
PrerequisiteLists are allocated and deallocated correctly when parsing the INI files, but they are not freed when owning ExtData object is destroyed. This way, at least one DVC is leaked for each TechnoType, plus the memory the DVC allocated.

The fix is using unique_ptr, which deletes the allocated DVCs when the ExtData is destroyed.

This - and the previous conversion to std::vector - makes maintaining the list easier. Superfluous items can just be erased.