Simplified a few bits
idxNewSWType was assigned GetNewTypeIndex(), or computed it manually. Using this index to resolve it to a NewSWType if it was above FIRST_SW_INDEX is what GetNewSWType() does.

IsTypeRedirected() implies it is an original SW type. Removing this check will let the original SW types and new SW types through now also, but if they aren't redirected, they will be caught in the GetNewSWType() method. If they are new types, they don't handle the Nuke, if they aren't the nuke.