Template Update 34: Split Template.h into two files.
TemplateDef.h contains the implementations for the reading and parsing methods. This only has to be included in cpp files that read INI data (usually the Body.cpp files).

TemplateDef.h should not be included in any Body.h or any file that is included in many other cpp files not related to reading INI data.

The functions extracted into the TemplateDef.h have been made inline. This is because the definitions file is included in more than one Body.cpp and thus the definitions would be there multiple times. Multiple definitions of inline methods don't violate the one-definition-rule.