Spotlight changes: Deallocator 'mismatch' and invalidation problems.
Spotlights are allocated using the game's heap, and they have a virtual destructor which frees using the game's heap. Thus 'delete' works here, but to make it explicit it has been changed to GameDelete.

Calling the BuildingLightClass destructor erases the item from the hash_map (line 88), and thus it is invalidated and shouldn't be erased again (line 99).

The other way around: First calling erase, then using the 'second' member of the invalidated iterator to call the destructor (line 123).