Changed spotlight from being managed by a special hash_map to being managed by TechnoExt.
Note that the spotlight constructor can be called with a null pointer argument (the vanilla game does this), so one can't assume there is a TechnoExt there. The previous code didn't acknowledge this and used a null pointer as key into the map, for every spotlight created without an owner.

This code might have changed behavior. Not sure.