Related to #895418: Sight > 10
Three functions have been hooked heavily. Loops that counted from a given offset in the CellSpread table to a predetermined endpoint have been replaced by CellSpreadEnumerators. The enumerator is placed as a pointer to a heap-allocated object on the stack of the methods, reusing memory.

Using larger Sight values will result in an octagonal shaped field of view. That is how CellSpread logic works.