Related to #896027: Ivan bombs can damage Ivan if Cloakable
Cloaking units disappear from the battlefield and their pointer are broadcast to all other objects. In this process, the BombListClass clears the Bombs' owners, ignoring the flag that distinguishes between disappearing and being removed completely.

This fix skips the call to the BombListClass method if the pointer is not announced as being permanently unavailable.