Related to #895990: CellSpread affecting buildings larger than 1x1 multiple times
[Warhead]CellSpread.MaxAffect= (integer, defaults to -1 (infinite))
The positive number of times an object can be affected at most when hit by a warhead using CellSpread. Buildings can take multiple hits, potentially one for each cell of their foundation. Set this to 1 to make them hit not more than once. If an object is hit more often than it should be, the hits closest to the impact site are used.