Related to #1283638: Ivans cannot enter Grinders
Ivan now gets the enter cursor over allied Grinders by default. If the Force Fire keys are pressed, the IvanBomb cursor is used (as it previously was).

This fix contains a bad hack, because otherwise it wouldn't be easy to return the correct value: There's a static byte array that serves as our dummy function end. It contains code to clean up the stack and to return the appropriate value.