Flying Aircraft Carriers
Spawns will be put in the air and not on the ground if the spawning unit is flying or hovering. Spawned units will move to the cell the spawner resides on, because they would lose their destination when it is in air.

For best effects, the FlightLevel of spawns should be higher than the FlightLevel of the spawning unit.

If the spawner unit is moving, it is possible that the spawns lower on the wrong cell and then gain altitude again. Maybe this can be made better, but... meh.