Draining only the affected Power Plant
[TechnoType]Drain.Local= (boolean, defaults to no)
Whether this unit drains only the local power from its target building, or whether this building does not bring down the entire power of the owning player. If any of the draining unit or the building drained have this set to yes, only the local power is drained. If both have this set to no, then the victim player's entire power is drained.

If a disk has Drain.Local=yes, it will only drain the local power of any power plant.
If a building has Drain.Local=yes, then any disk will only drain the local power.