Specify drain amount for local power drain
[TechnoType]Drain.Amount= (integer, defaults to 0)
The amount of power being drained from a power plant of the victim player. Requires Drain.Local on either the draining unit or the drained building.

If positive, this is the maximum amount of power drained. If negative, this is the minimum amount of power drained. The default value is the victim buildings Power=, then the drainer's Drain.Amount is applied, and finally the drained building's Drain.Amount.

For example: draining a Power Plant with Power=500 and Drain.Amount=0 using a Floating Disk. With Drain.Amount=-1000 on the disk, it will drain 1000, with Drain.Amount=0 it will drain the 500 the building produces, and with Drain.Amount=200 on the disk it will drain 200.

If the building would have Drain.Amount=-600, it would never drain less than 600. If the building had Drain.Amount=800, it would never drain more than 800.