Allow Tech and Occupiable Structures to be returned
Only buildings that initially belonged to a MultiplayPassive house are eligible for this behavior. In single player mode these settings are ignored.

[General]ReturnStructures= (boolean, defaults to no)
Whether structures will be returned to the civilian side when a player is defeated or gives up.

[BuildingType]Returnable= (boolean, defaults to [General]ReturnStructures)
Whether this building type will be returned to the civilian side when the owning player is defeated or gives up.