Reworked the super weapon update code.
Extracted some parts as a free function, which is bad style. The incorporation of the house's power state and the super weapon's disableableability (hey, this word is missing in my dictionary) has been included there as well. There were slight changes to the power checks, as well as some stylistic changes.