Hooked another super weapon checking function.
There are two function checking super weapons: one grants all the ones the player doesn't have yet and that just became available, the other puts the ones the player owns on hold in low power situation, resumes charging or removes the super weapon in case there are no more buildings that provide that super weapon.

This should fix the minor bug that super weapons requiring an operator start charging during the buildup animation, and then go on hold when it finishes.