More refactoring and a minor change.
Extracted a method that adds the experience to a unit, added some checks. Instead of multiplying the cost with all applicable factors, the one factor is maintained and multiplied by cost when being applied.

Changed MindControlledByHouse to the current owner (MC-self modifier), because the former might not be set when a unit is captured by another one. Now it uses the Owner field (in the lambda), which is the appropriate house in all cases of mind-control.