Swapped mind-control self and victim calculation.
It did not work as documented. The victim got the victim multiplier, but the controller got the self multiplier times the victim multiplier, instead of just the self multiplier. Thus the comment about splitting experience by setting both multipliers to sum up to 100% would be wrong.

Now, the controller gets only the self multiplier, the mind-control victim only gets the victim multiplier.