Veterancy from Spawns (Aircraft Carriers and Destroyers)
Spawners can now get experience from their spawns' kills. For this to work, both
the spawn and the spawner must be Trainable=yes. The following two tags go on
the unit that spawns, not the spawns themselves.

[TechnoType]Experience.SpawnOwnerModifier= (float, defaults to 0%)
The experience the unit owning a spawn gains when the spawn kills a unit. This
tag has to be set on the spawner, not the spawn.

[TechnoType]Experience.SpawnModifier= (float, defaults to 100%)
The experience the a spawn of this unit gains when the spawn kills a unit. This
tag has to be set on the spawner, not the spawn.

If you want to split the experience between spawner and spawns, you can make
these values sum up to 100%. This is not required, though.

If a spawner is mind-controlled, both the spawner's and the spawn's experience
is multiplied by Experience.MindControlVictimModifier.