Fix for Sight issues
Jump back in earlier, because this hook missed to set a stack variable to the same value as AX.

The pRel change is not the fix. The game now sets the value to the contents of EDX, so this is just another way to do the same as it did earlier.