Customizable campaign and multiplayer score screens
The campaign score screen consists of four elements, a background, a transition piece, an animation, and a palette file for all three.

[Side]CampaignScore.Background= (shp filename, including extension)
A 632x568 image used as the background. Uses the palette defined as CampaignScore.Palette. Defaults to ASCRBKMD.SHP, SSCRBKMD.SHP or SYCRBKMD.SHP for sides 1, 2 and all others respectively.

[Side]CampaignScore.Transition= (shp filename, including extension)
An image used as a transition from the looping animation to the static background image. Played backwards for the opposite case. Uses the palette defined as CampaignScore.Palette. Defaults to ASCRTMD.SHP, SSCRTMD.SHP or SYCRTMD.SHP for sides 1, 2 and all others respectively.

[Side]CampaignScore.Animation= (shp filename, including extension)
An image played as a loop after the introductory transition until the dialog is closed. The exact looping point is not known. Uses the palette defined as CampaignScore.Palette. Defaults to ASCRAMD.SHP, SSCRAMD.SHP or SYCRAMD.SHP for sides 1, 2 and all others respectively.

[Side]CampaignScore.Palette= (pal filename, including extension)
The palette to draw all animations on the campaign score screen. Defaults to ASCORE.PAL, SSCORE.PAL or YSCORE.PAL for sides 1, 2 and all others respectively.

The multiplayer score screen consists of a background drawn with a special palette, and ten pcx bar images.

[Side]MultiplayerScore.Background= (shp filename, including extension)
A 632x568 image used as the background. Uses the palette defined as MultiplayerScore.Palette. Defaults to MPASCRNL.SHP, MPSSCRNL.SHP or MPYSCRNL.SHP for sides 1, 2 and all others respectively.

[Side]MultiplayerScore.Palette= (pal filename, including extension)
The palette to draw the background of the multiplay score screen. Defaults to MPASCRN.PAL, MPSSCRN.PAL or MPYSCRN.PAL for sides 1, 2 and all others respectively.

[Side]MultiplayerScore.Bars= (pcx filename, including extension)
The filename used as a pattern for the ten bars of 440x36 on the multiplayer score screen, two captions and up to eight players. Can contain ~~, which are replaced with a number from 01 to 10. Defaults to mpascrnlbar~~.pcx, mpsscrnlbar~~.pcx or mpyscrnlbar~~.pcx for sides 1, 2 and all others respectively.