Customizable target area for Splits and Airburst
[Projectile]AroundTarget= (boolean, defaults to Splits)
Whether a projectile with Splits=yes or Airburst=yes should use the area around the original target to look for new targets for each cluster. If enabled, the clusters will continue their way to the originally intended target. Otherwise, the clusters will use the area where the projectile split up.