Graphical Text banner per side
[Side]GraphicalText.Image= (shp filename including extension, defaults to GRFXTXT.SHP)
The graphic file used as overlay to show whether a mission or match has been won or lost. The image is centered in the viewport.

[Side]GraphicalText.Palette= (pal filename including extension, defaults to GRFXTXT.PAL)
The palette the graphical text image uses. The banner is actually rendered in the theater's palette. This palette is used to find the best match for the colors in the theater pal.

The shp should have at least 4 frames, campaign won and lost, and multiplayer won and lost.