Side-specific score screen music themes
[Side]CampaignScore.UnderParTheme= (theme id, defaults to SCORE)
The theme playing for this side when a player finishes a campaign mission faster than the defined par time.

[Side]CampaignScore.OverParTheme= (theme id, defaults to SCORE)
The theme playing for this side when a player finishes a campaign mission slower than the defined par time.

[Side]MultiplayerScore.WinTheme= (theme id, defaults to SCORE)
The theme playing for this side when a player is victorious in a multiplayer match.

[Side]MultiplayerScore.LoseTheme= (theme id, defaults to SCORE)
The theme playing for this side when a player is defeated in a multiplayer match.