Customize the dialog background for each side
[Side]DialogBackground.Image= (shp filename, including extension)
The shp file used as background for dialog boxes for this side. Should be 452x326. Defaults to PUDLGBGA.SHP, PUDLGBGS.SHP, and PUDLGBGY.SHP for sides 1, 2 and all others respectively. Requires DialogBackground.Palette= to be set.

[Side]DialogBackground.Palette= (pal filename, including extension)
The palette used to draw the background of dialog boxes for this side. Defaults to DIALOG.PAL for sides 1 and 2, to DIALOG.PAL otherwise. Requires DialogBackground.Image= to be set.