Custom campaign loading screen music
[Basic]LoadingTheme= (theme id, defaults to LOADING)
The theme playing when loading this mission. Use "none" to disable the loading screen music.

No solution yet for multiplayer loading screens. The music starts to play before the loading screen is drawn and before the side index a player is on is available. When the information is available, we are in the middle of loading already (not literally). It should be possible to pre-calculate this, but I'll leave that for later.