Extension update: Got rid of the non-const Container::Find().
It made no sense to have a const version return a const ValueType. The constness is for the returned pointer, which ValueType always is, thus the const is just stripped away immediately. Also, the const_cast to non-const was bad design, which is improved by this.

Because the original non-const version did not return anything that could be used to change the Container, it is made const now.