Implemented AresPCXFile
This class encapsulates a PCX filename and provides a method to load the Surface efficiently. It will try to access the data and, in case it's not available, load the file only once. When the data is available, it will not load the file and thus not re-load a file that's loaded already.