Call the ExtData destructors later.
This is done because the destructor might call other methods which are hooked. The objects are still valid at this point, as no members have been torn down and the base class constructor has not yet been called.

If the original destructors call AnnounceInvalidPointer, the ExtData destructor is called only after that.

The item might be removed from the static Array. It is not for HouseClass, because ReversedByHouses needs it to be set.