Refactored the loop
FirstTarget was always assigned pData->PrismForwarding.SupportTarget. If LastTarget->SupportTarget was null, so was FirstTarget (the default was nullptr) -> return 0. Otherwise, FirstTarget was set to NextTarget, which was LastTarget->SupportTarget.

The alliance check against FirstTarget's owner has been moved up, as its result can't change during the loop.

LastTarget is resolved in a while loop that has been extracted. This gets rid of NextTarget. The inner check has been moved out below, as it has to be run only once for the last tower that doesn't support any other tower. This tower always exists. LastTarget is initialized non-empty and its successor exists, otherwise the FirstTarget check would fail. Thus there is a last element.

It would have been possible to get rid of LastTarget also and just reuse FirstTarget, but this might be easier to read.