Commented out BuildingExt::cPrismForwarding::AnnounceInvalidPointer.
The way it worked was slow and improper. It cast the pointer to a BuildingClass, checked whether it had an ExtData attached, did the checking, then removed the Building pointed to from the network, This was done for each building on the map, all on the same pointer.

As there have been reported no errors, the debugging lines have been removed.

Also included: some minor refactoring.