Customizable DamageParticleSystems
Ares adds two new tags that allow particle systems other than the ones with BehavesLike=Smoke|Spark.

[TechnoType]DamageSmokeParticleSystems= (list of ParticleSystems, defaults to all DamageParticleSystems with BehavesLike=Smoke set)
Defines a list of ParticleSystems to randomly spawn from when an object is damaged into yellow or red health. You have to use this if you want to use ParticleSystems with BehavesLike= set to other values than Smoke.

[TechnoType]DamageSparkParticleSystems= (list of ParticleSystems, defaults to all DamageParticleSystems with BehavesLike=Spark set)
Defines a list of ParticleSystems to randomly spawn from when an object is in yellow or red health. You have to use this if you want to use ParticleSystems with BehavesLike= set to other values than Spark.