Mod info tags
These are at the moment purely diagnostic - not shown anywhere, nor used anywhere. They can be defined in uimd.ini, which is mod-controlled. Mods can then specify their name and their version, and this will be printed in the debug.log.

[VersionInfo]Name= (string, 0x3F chars)
The name of the mod in unspecified format.

[VersionInfo]Version= (string, 0x3F chars)
The version of the mod. Unspecified format, not just numbers.

An identifier is created from these using the game's Checksummer. This is reserved for future use to identify a mod by name and version using a number comparison (collisions are allowed, but that shouldn't happen considering the number of mods and their release cycles compared to the nearly endless possible values an int can store).