Mass actions
An overly complicated system to automatically call the most appropriate method for pointer invalidation or clearing global state. Also keep calls to these methods in sync, so updating one place is enough and the order can never get messed up.

There are dispatcher classes that use type inference and decltype to select the most appropriate method for each operation. If no callable method can be found, the type is skipped with no overhead. Just implement a method that a dispatcher accepts, and it will be picked automatically.

The types to dispatch do not have to be in the same class hierarchy. The system is completely agnostic to the types' implementation.