Scorch and Flamer tags reimplemented
Tiberian Sun had Scorch and Flamer for animations, two tags that spawned fire animations defined by SmallFire and LargeFire. This has been reimplemented as it worked in Tiberian Sun, with the exception that it would not crash if SmallFire or LargeFire aren't defined, because some YR animations have Scorch=yes set while the animations are commented out.