Deliver.Owner for UnitDelivery SWs
[SuperWeapon]Deliver.Owner= (enum - invoker, civilian, special, neutral, defaults to invoker)
The house the placed units will belong to. Invoker is the player who fired the super weapon; Civilian is the first house of the side named 'Civilian'; Special is the house named 'Special'; Neutral is the house named 'Neutral'.

This is a new approach to use the Valueable mechanism for reading special values instead of using the ValueableEnum template with a Valueable<int> in the background. This is possible because C++11 enum classes are real types that can be used for overloading and specialization, while the old mechanism of unscoped enums in classes was just typedeffing something (and typedefs are the same as the original type when it comes to overloading).