AI support for TogglePower
This makes the AI a bit smarter when it comes to coping with power outages because of missing power plants. At the moment it is hard-coded to on, new tags to enable it will come in later.

Each building is evaluated, and base defenses are tried to be held online if possible. This is done by multiplying the negative Power= by 2 for buildings providing super weapons, and by 1.5 by non-base-defenses. Higher number increases the chance of the building being turned off, lower numbers mean buildings are turned back on earlier.

Players being drained or suffering from a power blackout triggered by a spy or Force Shield are exempt, as well as any human players.