Updated the SW LaunchSite code.
A virtual function GetLaunchSiteRange() now returns the general range of the SW, which can depend on the firing building. If pBuilding is nullptr, a fake range can be returned to always pass the checks in HasLaunchSite().

SW.MaximumRange and SW.MinimumRange are now disabled if < 0.0.

Code now uses the standard algorithms instead of poorly hand-coded iterators.