Several changes to the UnitDelivery SW
Support WaterBound structures again. Also change the coordinates so the current object to place appears to be centered on the actual target coordinates.

Also changed the extent variables to short, to not have conversions there, and changed the Put direction to be unsigned. An abstract_cast has been changed to static_cast, because the type is known without check.