Manually destroy VoxClass instances
Added a cast to VoxClass2, which then calls the VoxClass2 destructor, when in turn calls the original VoxClass destructor. This has to be done, because VoxClass's destructor isn't virtual.

'delete' matches the VoxClass2 creation using 'new'.

The array is iterated in reverse order, because each destructor removes the item from the list. Starting at the front just requires all items to be moved down for no reason. Not that it isn't a waste already calling FindItemIndex for all of them...