Extracted the part that does the basic extension initialization into a method called EnsureConstanted. It's only applied once, and every consecutive call then does nothing any more.
This changes behavior, because now the Initialized field is updated after calling InitializeConstants. Objects are 'constanted' right after construction (in Container::FindOrAllocate). Previously, when an ini file was read later, it would be called again for each and every ext object because Initialized was not updated, and the switch in LoadFromINI stepped in too early.