Buildings stopping to fire when occupant is killed by Pass Through logic
The current firing occupant number wasn't updated in case an occupant died. The index could became invalid, if the next occupant to fire was the last occupant. In this case, any removal from list made this index invalid. The building then wouldn't auto-target and not give an attack cursor. Ejecting and re-entering fixed this, as leaving the building resets the index.