Secret Labs: Moved over bonus assignment to BuildingExt.
This fixes the unreported problem of only one SecretLab=yes building of a type being assigned anything from SecretLab.PossibleBoons= if SecretLab.GenerateOnCapture=no (default). This was because the building type remembered whether it assigned anything to any building, and the next building of the same type would exit quickly without assigning anything.