HouseExt::PickIonCannonTarget, a replacement for the game function.
Uses IsIonCannonEligibleTarget and GetIonCannonValue.

Supports an optional Super Weapon type, which is used to check ranges, designators and such. It is also possible to pass an enumeration value to decide what to do with cloaked units.

pOwner is optional. If not set, all houses are eligible. Otherwise only the house passed is allowed.