Multiple AuxBuildings and negative aux buildings
[SuperWeapon]SW.AuxBuildings= (list of BuildingTypes, defaults to none)
Any of these buildings need to be owned by the player to be able to use this super weapon. Use this instead of AuxBuilding if you want to have multiple AuxBuildings,

[SuperWeapon]SW.NegBuildings= (list of BuildingTypes, defaults to none)
The super weapon becomes unavailable if a player owns any of these buildings.

This is not fully hooked up yet. In several checks the new tags aren't accounted for.