Moved SW.RequiredHouses, SW.ForbiddenHouses, SW.AuxBuildings and SW.NegBuildings checks
It is now in the same place like the AuxBuilding check, which means it now affects all super weapons, be they SuperWeapon or SuperWeapon2 on buildings or upgrades. This means, certain buildings are now considered not to have super weapons.

This does not affect the way buildings are made unbuildable because of disabled super weapons. This is still controlled by SuperWeapon unconditionally. Thus, even if the super weapon isn't available, the building would become unbuildable.