EVA and Text messages when building is captured
[BuildingType]LostEvaEvent= (EVA event, defaults to EVA_TechBuildingLost)
The EVA message to play when a building of this type is captured from the player. Requires NeedsEngineer=yes.

[BuildingType]Message.Capture= (csf label, defaults to none)
The text being printed when a building of this type is captured by the player. Requires NeedsEngineer=yes.

[BuildingType]Message.Lost= (csf label, defaults to none)
The text being printed when a building of this type is captured by the enemy. Requires NeedsEngineer=yes.