Broken PrismSupportModifier fix
The PrismSupportModifier bugfix itself contained a bug that caused the modifier to be rounded down to a multiple of 100 when an INI was read and didn't contain the tag. This was because integer division was used instead of floating point division.

The bug has been fixed so now the current PrismSupportModifier is divided by 100.0, which is then multiplied by 100.0 again to retrieve the correct value.

Thanks to Iran for reporting this.