Added ExtData overrides to save and load data
The actual code to read and write the data is still missing, so at this moment this is just wasting space.

This is implemented using a templated helper instance method Serialize(), as this allows to share the implementation between reading and writing and also to access the private data.

WARNINGS:
*If new members are added, see that they are also mentioned here. (If they need to be saved)*

*Do not mess with the order just for the sake of ordering them. This would break previous save games.*

*Removing or adding members breaks older save games.*