AresStreamObject with several specializations
These are surrogates that know how to load and save objects that do not know how to do that themselves and aren't POD types without pointers. If there are pointers involved, they need to be swizzled, and thus copying the memory alone wouldn't help.

Supports the following types:
- VectorClass<T>
- DynamicVectorClass<T>
- TypeList<T>
- CounterClass
- SHPStruct*
- RocketStruct
- CameoDataStruct
- std::bitset<Size>
- std::string
- std::vector
- Valueable<T> and all derived classes
- ValueableVector<T> and all derived classes
- Promotable<T>
- ParadropPlane
- all classes that have the has_loadsave_members trait enabled

This system a bit older already, and there are better ways to implement this. But that will be another commit.