Support persisting and restoring objects
All objects pointed to by a unique_ptr member must be owning the instance pointed to, thus they will now store it also when the pointer is saved. When loading, the object is restored into the unique_ptr.

Swizzling is supported, so ordinary pointers to an object managed by a unique_ptr are possible.