Make savegames depend on the Ares version that created them and the Mod ID
Savegames that do not suffice these will not be shown in the list.

From now on, whenever a new tag or member is added or removed, or any of the save and load methods is changed to process fields in a different order, the version number has to be changed, otherwise the old savegames will still appear in the list, and might cause the game to crash if loaded.

If you need to update the version number without any change of binary layout and want to keep old savegames working, you'll have to set SAVEGAME_MAGIC to the constant calculated from the old version number.