Switch between calling upper and lower case savegame methods on types automatically, falling back to the surrogate approach if that fails.
It should be easier to use as Ares picks up the load and save methods automatically, and no work has to be done to enable this behavior first. Also, it's easier to support class templates.

This gets rid of has_loadsave_members, ENABLE_ARES_PERSISTENCE, the second type parameter on AresStreamObject used for enable_if, as well as all custom AresStreamObject specializations for class templates that would just call the load and save methods.