Related to #1395457: Paradrop and cloak issue
Units having a parachute are now unable to cloak or be cloaked, just like in the original game. Originally, this was a consequence not of the parachute but of the fact the unit's height was above 0, which always prevented cloaking.

With Ares supporting cloak even for units not on the ground, this fix should come as close to the original as needed.