Changed the way ExtData items are removed
- Move ExtData pointers out of the map before erasing them on Remove(). It will be destroyed when the function completes.

- Swap the entire map with an empty one on Clear(), and then have the data destroyed when the function completes.

Consequences:
a) An ExtData object will not be notified of the objects it itself invalidates during destruction. This shouldn't be a problem, because the destructor is responsible for what it does anyhow.

erase() first destroyed the unique_ptr object in the map, which in turn is required to move out the pointer into a local, set itself to null and then invoke the destructor on the moved out pointer. This means that while the erase was in progress, the Items map contained a null pointer, which had to be checked for. Now, the pointer is moved out and kept alive, the destructor called in erase does nothing on the null pointer and thus there cannot be any invalidation during erase() any longer, and the moved out pointer is destroyed when the function finishes.

b) When using Clear(), no ExtData of a type is notified of the objects other ExtData instances invalidate. This should not really matter, because Clear() is called only when everything is cleaned up, and destructors shouldn't mess with the members in such a way if the data is not owned by the ExtData. Also, if Clear() does have to remove items, that means there is a code path that was left unchecked. The normal way is that each object is removed separately. If that isn't true, there's a more severe problem anyhow. In other words: Clear() is not supposed to do anything.

c) An object being erased cannot be looked up any more. Find() will return a null pointer. This is not new though, as the unique_ptr is set to a null pointer before destruction anyhow (as explained above). If this is a problem, this means some logic still relied on objects being destroyed at that very moment.

d) The for loop for calling InvalidatePointer on all ExtData objects can now get rid of the null pointer check, which should save several infinitely small amount of time a large number of times. No idea how long that is.