Added a Name member to Container<T>
This is initialized in the constructor, which has been made explicit. The Name value is used instead of the typeid values, which saves about 25kb in the built dll, as these were the only uses of typeid.

This unified two types used along with typeid: the base_type and the type of *this, which was the Container<T> name and wasn't that appropriate in at least one location anyhow.