Updating SupportingPrisms like this is an error.
When a tower fires, it is removed from the network before the supported tower fires. This means for each support tower, the value is decremented on the support tower, which messes with drawing lasers in different intensities. Thus, master towers' laser sometimes erroneously renders as if it wasn't supported.

Example: Have a chain of three towers. The last one has no senders, thus its value is 0. Its sent-to towers' values in the chain are 1 and 2. The first one fires, and removes itself, so the middle one's value is changed from 1 to 0 before it fires. Then it fires and removes itself, decrementing the final tower to 1.

The correct way would be to recalculate this value (and all the following ones in the network) so the chain length is maintained even if a middle tower is removed, but not when it fires. As this is only a visual problem, this issue is ignored.