Reworked Debug::DumpObj, provided a template for inferring size automatically.
Both paths are now unified, and printing full lines and printing the last line are now handled by the same loop. This also fixes the minor issue that if the size was divisible by 16 there was an extra line appended with no values at all.

The method now accepts pointers to const void, which makes casts unnecessary and also accepts const data.

The structure has changed a bit, and the empty lines have been removed in favor or a divider line below the header.