Fixed a regression in map reveal
When translating, I eliminated two calls to both sqrt and F2I, and then compared to the square of the range. This was not right, as F2I would floor all roots and thus allow all distances less than spread+1.

The fix is to use sqr(spread+1) and a less than comparison. sqrt(spread_limit_sqr) is exactly spread+1, larger than spread, and thus too large to be valid. Any squared distance less than spread_limit_sqr will result in a root strictly less than spread+1, which in terms of integers (with their built-in floor) is less than or equal to spread. This is what we want.