Replaced NukeSilo search by the new LaunchSite mechanism. This supports SuperWeapons=.
The logic changed a bit: Instead of iterating all BuildingTypes, looking for the ones with NukeSilo=yes and then finding a building of that type, it now looks at each building whether it is a valid launch site.

The original game function supports a mechanism to select buildings by location, but nuke code never made use of it. There are changes, though: power requirements are now checked, the silo has to have health and must be alive. The selected silo is the one built first, not the first one of the first type with NukeSilo=yes.