Fix full-map Sonar Pulse
This has been broken for some time now, because <0,<0 is considered an empty range and thus defaults to 10,-1. Thus, full map sonar could only be triggered by using ranges of <0,>0 explicitly.

The fix is to change SW.Range parsing to not parse single values as w,-1 but as w,0. Thus, the range does not count as empty. Fixed a spot (on Lightning Storm) where a width of 0 would trigger rectangle mode.

Also support going back from full map detection to ranged detection, if the range is changed in a map file.