Changed the defaults for targeting types Self and ParaDrop.
Self does only look at buildings of the owner, and there's no way to change this. None is faster than Buildings, because the check can be skipped altogether. Building could be wrong, as the center cell of a building might not even be occupied by a building.

ParaDrop requiring enemy objects only if there is an object at all does not help much. And it will not work if the paradrop is targeting the owning player's base.