Global options for DamageDelay
[General]Degrade.Percentage= (double - percentage, defaults to [AudioVisual]ConditionYellow)
The health above which this building will continue to degrade during low power situations. Buildings with health equal to or below this will not degrade further.

[General]Degrade.AmountNormal= (int - hitpoints, defaults to 0)
The damage caused to buildings with Power equal to or greater than 0 in low power situations.

[General]Degrade.AmountConsumer= (int - hitpoints, defaults to 1)
The damage caused to buildings with Power less than 0 in low power situations.