DamageDelay settings on BuildingTypes
[BuildingType]Degrade.Percentage= (double - percentage, [General]Degrade.Percentage)
The percentage of health this building will degrade to in low power situations. A value of 1.0 will not make it degrade, a value of 0.0 will make it degrade till complete destruction.

[BuildingType]Degrade.Amount= (int - hitpoints, defaults to the global rules)
The amount of damage this building will receive each interval of DamageDelay in low power situations. A value of 0 will make this building not degrade.