Made Firestorm immolation anims customizable
[AudioVisual]FirestormGroundAnim= (AnimationType, defaults to FSGRND)
The animation to display when an active Firestorm Wall destroys an object on the ground.

[AudioVisual]FirestormAirAnim= (AnimationType, defaults to FSAIR)
The animation to display when an active Firestorm Wall destroys an object in the air.

Added a new function on RulesExt:ExtData that is called once after the type data has been parsed. This is required because otherwise the animations couldn't default to a value, because none are loaded. Automatically allocating is not a good option.