Reworked Firestorm Wall idle animations to behave as in Tiberian Sun
Changes to the previous Ares versions are that the animation is not delete'd every 8 frames even if not finished and that the animation now has some other properties: offset is now a bit different (768 -> 740), zAdjust is now -10 and the flags have 0x4 set now.

Another internal change is the storage location of the created animation. It is now put onto the wall building in slot SpecialTwo, while previously an extra field was used, which might not be named correctly: It is an abstract pointer for sure, but unused; it exists in TS, where it is not used for Firestorm animations (in fact, it is unused there, too).