Changed a huge part of Firestorm's behavior.
As now active Firestorm Wall sections can take damage in the form of reduced remaining charge time, Firestorm Walls can now be valid targets. Thus, the GetFireError() hook is no longer useful.

Also, bullets don't need to be made to hit the Firestorm Wall artificially, as now Inviso=no bullets are caught in BulletClass::Update() and Inviso=yes bullets are redirected in BulletClass::MoveTo().

Removing this code should have a huge impact on performance when a Firestorm is activated, as GetFireError() is called all the time to probe a plethora of objects even if far away, not just when actually firing.